split the catalog and the filter
