+++ src/pages/ui/catalog-filter/index.ts
@@ -1,0 +1,1 @@
+export { CatalogWithFilter } from './catalog.filter';
