+++ src/widgets/catalog/index.ts
@@ -1,1 +1,1 @@
-export { Catalog } from './ui';
+export { Catalog } from './catalog';
