--- src/widgets/catalog/ui/index.ts
@@ -1,1 +1,0 @@
-export { Catalog } from './catalog';
