--- src/widgets/catalog/ui/product-list/index.ts
@@ -1,1 +1,0 @@
-export { ProductList } from './product-list';
