+++ src/widgets/filter/index.ts
@@ -1,0 +1,1 @@
+export { Filter } from './ui';
